// src/features/artifact/lib/enums.ts
import { z } from 'zod';
export enum ArtifactType {
  Active_AdvanceTimeByMinutes = 'Active_AdvanceTimeByMinutes',
  Active_FieldHospital = 'Active_FieldHospital',
  Active_GetMoneyCache = 'Active_GetMoneyCache',
  Active_GetResourceCache = 'Active_GetResourceCache',
  Active_GetArmyCache = 'Active_GetArmyCache',
  Active_InspireMercenaries = 'Active_InspireMercenaries',
  Active_MarketRaid = 'Active_MarketRaid',
  Active_ReduceArmiesNeededToConquerTerritory = 'Active_ReduceArmiesNeededToConquerTerritory',
  Active_RevealTerritories = 'Active_RevealTerritories',
  Active_SuperchargeArmyCampByMinutes = 'Active_SuperchargeArmyCampByMinutes',
  Active_SuperchargeMineByMinutes = 'Active_SuperchargeMineByMinutes',
  Active_UpgradeArmyCampForDiscount = 'Active_UpgradeArmyCampForDiscount',
  Active_UpgradeHospitalForDiscount = 'Active_UpgradeHospitalForDiscount',
  Active_UpgradeMineForDiscount = 'Active_UpgradeMineForDiscount',
  Active_TripleStrike = 'Active_TripleStrike',
  Active_QuadrupleStrike = 'Active_QuadrupleStrike',
  AllCachesIncreased = 'AllCachesIncreased',
  AlloySellValues = 'AlloySellValues',
  ArmyCampUpgradeDiscount = 'ArmyCampUpgradeDiscount',
  ArmyCampProduction = 'ArmyCampProduction',
  CacheArmiesIncreased = 'CacheArmiesIncreased',
  CacheMoneyIncreased = 'CacheMoneyIncreased',
  CacheResourcesIncreased = 'CacheResourcesIncreased',
  CraftDouble = 'CraftDouble',
  CrafterIngredientsReducedBy = 'CrafterIngredientsReducedBy',
  CrafterSpeedIncreasedBy = 'CrafterSpeedIncreasedBy',
  DraftsIncreasedBy = 'DraftsIncreasedBy',
  HospitalUpgradeDiscount = 'HospitalUpgradeDiscount',
  HospitalSaveAdditional = 'HospitalSaveAdditional',
  ItemSellValues = 'ItemSellValues',
  MaxIdleTime = 'MaxIdleTime',
  MercenaryDiscount = 'MercenaryDiscount',
  MineUpgradeDiscount = 'MineUpgradeDiscount',
  MoneyFromBonuses = 'MoneyFromBonuses',
  MoneyFromTerritories = 'MoneyFromTerritories',
  OreSellValues = 'OreSellValues',
  MineProduction = 'MineProduction',
  SmeltDouble = 'SmeltDouble',
  SmelterIngredientsReduced = 'SmelterIngredientsReduced',
  SmelterSpeedIncreased = 'SmelterSpeedIncreased',
  TechDiscount = 'TechDiscount',
}

export const artifactTypeSchema = z.nativeEnum(ArtifactType);

export enum ArtifactRarity {
  Poor = 0,
  Common = 1,
  Uncommon = 2,
  Rare = 3,
  Epic = 4,
  Legendary = 5,
  Insane = 6,
}